+++ bevy_rapier2d/examples/contact_filter2.rs
@@ -1,4 +1,4 @@
-use bevy::prelude::*;
+use bevy::{ecs::system::SystemParam, prelude::*};
 use bevy_rapier2d::prelude::*;
 
 #[derive(PartialEq, Eq, Clone, Copy, Component)]
@@ -11,15 +11,15 @@ enum CustomFilterTag {
 // same user_data value.
 // Note that using collision groups would be a more efficient way of doing
 // this, but we use custom filters instead for demonstration purpose.
-struct SameUserDataFilter;
-impl<'a> PhysicsHooksWithQuery<&'a CustomFilterTag> for SameUserDataFilter {
-    fn filter_contact_pair(
-        &self,
-        context: PairFilterContextView,
-        tags: &Query<&'a CustomFilterTag>,
-    ) -> Option<SolverFlags> {
-        if tags.get(context.collider1()).ok().copied()
-            == tags.get(context.collider2()).ok().copied()
+#[derive(SystemParam)]
+struct SameUserDataFilter<'w, 's> {
+    tags: Query<'w, 's, &'static CustomFilterTag>,
+}
+
+impl BevyPhysicsHooks for SameUserDataFilter<'_, '_> {
+    fn filter_contact_pair(&self, context: PairFilterContextView) -> Option<SolverFlags> {
+        if self.tags.get(context.collider1()).ok().copied()
+            == self.tags.get(context.collider2()).ok().copied()
         {
             Some(SolverFlags::COMPUTE_IMPULSES)
         } else {
@@ -36,7 +36,7 @@ fn main() {
             0xFF as f32 / 255.0,
         )))
         .add_plugins(DefaultPlugins)
-        .add_plugin(RapierPhysicsPlugin::<&CustomFilterTag>::pixels_per_meter(
+        .add_plugin(RapierPhysicsPlugin::<SameUserDataFilter>::pixels_per_meter(
             100.0,
         ))
         .add_plugin(RapierDebugRenderPlugin::default())
@@ -56,10 +56,6 @@ pub fn setup_physics(mut commands: Commands) {
     /*
      * Ground
      */
-    commands.insert_resource(PhysicsHooksWithQueryResource(Box::new(
-        SameUserDataFilter {},
-    )));
-
     let ground_size = 100.0;
 
     commands.spawn((